Impl

+++ components/hlink_ac/hlink_ac.h
@@ -25,7 +25,7 @@ namespace esphome
       optional<esphome::climate::ClimateMode> mode;
       optional<esphome::climate::ClimateFanMode> fan_mode;
       optional<esphome::climate::ClimateSwingMode> swing_mode;
-      optional<uint64_t> device_sn;
+      optional<uint32_t> device_sn;
       bool ready()
       {
         return power_state.has_value() && current_temperature.has_value() && target_temperature.has_value() && mode.has_value() && fan_mode.has_value() && swing_mode.has_value();
@@ -74,7 +74,7 @@ namespace esphome
         INVALID
       };
       Status status;
-      uint64_t p_value;
+      uint32_t p_value;
       uint16_t checksum;
     };
